Average pooling clamped divisor should be done on all conditions where the kernel can go out of bounds (#4144)
In this pull request I added various E2E tests to cover edge cases on
the average pooling torch to linalg lowering algorithm. These new tests
uncovered various numerical issues that were addressed in this same PR.

One of the issues is the IREE test failure found in
https://github.com/llvm/torch-mlir/issues/4079 which triggered this
work.

Background:

In the most common case, the divisor for the average pooling is just the
product of kernel dimensions. But with padding, and ceil mode options,
some elements need to be discounted from the divisor computation. This
change fixes two components of this:

Fix the condition that determines if the divisor is just the product of
kernel dimensions or the clamped divisor comptutation.
Add missing isCountIncludePad logic divisor computation algorithm and
reversal of kernel/stride/padding parameters element order. Both were
missing from the first generalization change.

---------

Co-authored-by: Ivan Garcia <igarcia@vdi-ah2ddp-178.dhcp.mathworks.com>
Co-authored-by: ivangarcia44 <ivangarcia44>