Integrate LLVM at eda3e96b401a9b86132e39432e41e2000d1ab382 (#4208)
Update LLVM to
https://github.com/llvm/llvm-project/commit/eda3e96b401a9b86132e39432e41e2000d1ab382.

The `ToMemrefOp` is replaced with `ToBufferOp`, as per the changes made
here:
https://github.com/llvm/llvm-project/commit/8f91b108df21bee8ee64e88ffa7bdee230389800.

Also, `getBackwardSlice` now has a return value of type LogicalResult.
Changes made here:
https://github.com/llvm/llvm-project/commit/6a8dde04a07287f837bbabeb93e23e47af366d3d.

This commit also adds the folder for `tril_indices` and `triu_indices`
op for the case when the `row` and `col` values are 0.

---------

Signed-off-by: Vivek Khandelwal <vivekkhandelwal1424@gmail.com>