[Tosa] : Use output type for bias for creating tosa.conv (#4252)
For ConvolutionLayer initialized without bias, a zero tensor for bias is
created when converting to `tosa.conv2d` as the op always expects a bias
tensor. This zero tensor was always initialized to be `fp32`
irrespective of what the input/weights type were. This leads to a
validation error since `bias` type (fp32) didn't match with output of
conv (fp16) when the input/weight are of `fp16` type.