Bump the manylinux image used by build_linux_packages.sh (#4376)
The current manylinux image ships GCC 10, which has recently started
failing to compile torch-mlir (*), see failing CI job:
* https://github.com/llvm/torch-mlir-release/actions/runs/19329445043/job/55288631613).

This PR bumps `manylinux` to the next version which happens to bump
GCC from 10 to 14. This fixes the build issue.

While there may be narrower fixes, the image we use is also quite dated,
so updating it is a reasonable path.

(*) More specifically, `StablehloOptimizationPasses` inside the StableHLO
submodule.